payment + promotion updates 3

+++ api/models/promotion.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime
+from sqlalchemy import Column, Integer, String, Float, DateTime, DECIMAL
 from ..dependencies.database import Base
 
 class Promotion(Base):
@@ -6,5 +6,5 @@ class Promotion(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     code = Column(String(50), unique=True, nullable=False)
-    discount = Column(Float, nullable=False)
+    discount = Column(DECIMAL(5, 2), nullable=False)
     expiration_date = Column(DateTime, nullable=False)
